% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCounter.R
\name{getCounter}
\alias{getCounter}
\title{Downloading datasets from the siat.stat.uz}
\usage{
getCounter()
}
\value{
Returns a data frame
}
\description{
Downloads available datasets and related information from siat.stat.uz API
}
\section{Functions}{
\itemize{
\item \code{getCounter()}: Collects the number of the available datasets on the siat.stat.uz portal.

}}
\examples{
 \dontrun{
 getTable("2540")
 getMetaData("2540") }
}
\author{
Alisher Suyunov
}
