% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCounter.R, R/getMetaData.R
\name{getCounter}
\alias{getCounter}
\alias{getMetaData}
\title{Downloading datasets from the siat.stat.uz}
\usage{
getCounter()

getMetaData(id)
}
\arguments{
\item{id}{a numeric or character that represents a dataset identifier}
}
\value{
Returns a data frame

Returns a data frame
}
\description{
Downloads available datasets and related information from siat.stat.uz API

Downloads available datasets and related information from siat.stat.uz API
}
\section{Functions}{
\itemize{
\item \code{getCounter()}: Downloads the corresponding meta-data for specified dataset as a data frame.

\item \code{getMetaData()}: Downloads the corresponding meta-data for specified dataset as a data frame.

}}
\examples{
 \dontrun{
 getTable("2540")
 getMetaData("2540") }
 \dontrun{
 getTable("2540")
 getMetaData("2540") }
}
\author{
Alisher Suyunov
}
