% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getName.R
\name{getName}
\alias{getName}
\title{Downloading datasets from the siat.stat.uz}
\usage{
getName(dataset_id)
}
\arguments{
\item{dataset_id}{a numeric or character that represents a dataset identifier}
}
\value{
Returns a data frame
}
\description{
Downloads available datasets and related information from siat.stat.uz API
}
\section{Functions}{
\itemize{
\item \code{getName()}: Returns the name of the dataset under the given ID.

}}
\examples{
 \dontrun{
 getTable("2540")
 getMetaData("2540") }
}
\author{
Alisher Suyunov
}
