% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTable.R
\name{getTable}
\alias{getTable}
\title{Downloading datasets from the siat.stat.uz}
\usage{
getTable(id)
}
\arguments{
\item{id}{a numeric or character that represents a dataset identifier}
}
\value{
Returns a data frame
}
\description{
Downloads available datasets and related information from siat.stat.uz API
}
\section{Functions}{
\itemize{
\item \code{getTable()}: Downloads the specified dataset as a data frame.

}}
\examples{
 \dontrun{
 getTable("2540")
 getMetaData("2540") }
}
\author{
Alisher Suyunov
}
