% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_siat.R
\name{availableData}
\alias{availableData}
\title{Downloading datasets from the siat.stat.uz}
\usage{
availableData()
}
\value{
Returns a data frame
}
\description{
Downloads available datasets and related information from siat.stat.uz API
}
\section{Functions}{
\itemize{
\item \code{availableData()}: Obtains the list of available datasets from the siat.stat.uz

}}
\examples{
 \dontrun{
 getTable("2540")
 getMetaData("2540") }
}
\author{
Alisher Suyunov
}
